requirements.txt
https://www.djangoproject.com/download/1.8a1/tarball/
django-environ==0.3.0
icalendar==3.8.4
